lambda/requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-ore
pinecone
python-dotenv
serpapi